Melee Damage Standardization and Improvements
Many melee weapons have been tweaked to be more useful based on rarity, size, and purpose. Generally large-sized weapons like sabres, katanas, and Cutlass' deal 32 slash (the damage the machete already has), and huge-sized weapons should deal around 40. Weapons that are supposed to deal more structural damage have also had that increased somewhat, mostly due to the existence of the breaching hammer.
I made this because i was tired of seeing 15 slash on the plasteel arming sword (which is the same damage as a cybersun pen btw), and the 22 damage on the katana (making it effectively a larger, slower, combat knife).